requirements.txt
asgiref==3.2.10
boto3==1.14.51
botocore==1.17.51
Django==3.0
django-restframework==0.0.1
django-social-share==2.0.0
django-storages==1.10
djangorestframework==3.11.1
docutils==0.15.2
jmespath==0.10.0
pkg-resources==0.0.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
